feat(python): Allow empty `select`/`with_columns`/`groupby`
