+++ py-polars/polars/lazyframe/groupby.py
@@ -120,8 +120,6 @@ class LazyGroupBy:
             raise ValueError(
                 f"'aggs' argument should be one or multiple expressions, got: '{aggs}'."
             )
-        if aggs is None and not named_aggs:
-            raise ValueError("Expected at least one of 'aggs' or '**named_aggs'")
 
         exprs = parse_as_list_of_expressions(aggs, *more_aggs, **named_aggs)
